fixup! Turbopack css: treat `composes` and `@import` as urls when resolving
