+++ crates/turbopack-css/src/references/import.rs
@@ -24,16 +24,16 @@ use swc_core::{
 use turbo_tasks::{Value, ValueToString, Vc};
 use turbopack_core::{
     chunk::{ChunkableModuleReference, ChunkingContext},
-    issue::{IssueSeverity, IssueSource},
+    issue::IssueSource,
     reference::ModuleReference,
-    reference_type::UrlReferenceSubType,
+    reference_type::CssReferenceSubType,
     resolve::{origin::ResolveOrigin, parse::Request, ModuleResolveResult},
 };
-use turbopack_ecmascript::resolve::url_resolve;
 
 use crate::{
     chunk::CssImport,
     code_gen::{CodeGenerateable, CodeGeneration},
+    references::css_resolve,
 };
 
 #[turbo_tasks::value(into = "new", eq = "manual", serialization = "none")]
@@ -345,12 +345,11 @@ impl ImportAssetReference {
 impl ModuleReference for ImportAssetReference {
     #[turbo_tasks::function]
     fn resolve_reference(&self) -> Vc<ModuleResolveResult> {
-        url_resolve(
+        css_resolve(
             self.origin,
             self.request,
-            Value::new(UrlReferenceSubType::CssAtImport),
+            Value::new(CssReferenceSubType::AtImport),
             Some(self.issue_source),
-            IssueSeverity::Error.cell(),
         )
     }
 }
