+++ crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/another.css
@@ -1,0 +1,3 @@
+.bar {
+  background-color: green;
+}
