+++ crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/other.css
@@ -1,3 +1,5 @@
+@import "another.css";
+
 .foo {
   background-color: red;
 }
