+++ crates/turbopack-tests/tests/snapshot/css/relative-uri-import/output/79fb1_turbopack-tests_tests_snapshot_css_relative-uri-import_input_another_7d7e1c.css
@@ -1,0 +1,5 @@
+/* [project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/another.css [test] (css) */
+.bar {
+  background-color: green;
+}
+/*# sourceMappingURL=79fb1_turbopack-tests_tests_snapshot_css_relative-uri-import_input_another_7d7e1c.css.map*/
