--- crates/turbopack-tests/tests/snapshot/css/relative-uri-import/output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_3c4e5b._.css
@@ -1,19 +1,0 @@
-/* [project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/other.css [test] (css) */
-.foo {
-  background-color: red;
-}
-
-/* import([project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/other.css [test] (css)) */
-
-/* [project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/other.css [test] (css) */
-.foo {
-  background-color: red;
-}
-
-/* [project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/index.css [test] (css) */
-body {
-  background-color: blue;
-}
-
-
-/*# sourceMappingURL=crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_3c4e5b._.css.map*/
