--- crates/turbopack-tests/tests/snapshot/css/relative-uri-import/output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_5f5097.js
@@ -1,12 +1,0 @@
-(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
-    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_5f5097.js",
-    {},
-]);
-(globalThis.TURBOPACK_CHUNK_LISTS = globalThis.TURBOPACK_CHUNK_LISTS || []).push({
-  "path": "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_5f5097.js",
-  "chunks": [
-    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_3c4e5b._.css",
-    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_c2ad30.js"
-  ],
-  "source": "entry"
-});
