+++ crates/turbopack-tests/tests/snapshot/css/relative-uri-import/output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_8c26d0.js
@@ -1,0 +1,6 @@
+(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
+    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_8c26d0.js",
+    {},
+    {"otherChunks":[{"path":"output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_99b643._.css","included":["[project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/another.css [test] (css)","[project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/other.css [test] (css)","[project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/index.css [test] (css)"],"moduleChunks":["output/79fb1_turbopack-tests_tests_snapshot_css_relative-uri-import_input_another_7d7e1c.css","output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_other_7d7e1c.css","output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_7d7e1c.css","output/79fb1_turbopack-tests_tests_snapshot_css_relative-uri-import_input_another_7d7e1c.css","output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_other_7d7e1c.css"]},"output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_c2ad30.js"],"runtimeModuleIds":["[project]/crates/turbopack-tests/tests/snapshot/css/relative-uri-import/input/index.js [test] (ecmascript)"]}
+]);
+// Dummy runtime
