+++ crates/turbopack-tests/tests/snapshot/css/relative-uri-import/output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_a0e1a8.js
@@ -1,0 +1,12 @@
+(globalThis.TURBOPACK = globalThis.TURBOPACK || []).push([
+    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_a0e1a8.js",
+    {},
+]);
+(globalThis.TURBOPACK_CHUNK_LISTS = globalThis.TURBOPACK_CHUNK_LISTS || []).push({
+  "path": "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_a0e1a8.js",
+  "chunks": [
+    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_99b643._.css",
+    "output/crates_turbopack-tests_tests_snapshot_css_relative-uri-import_input_index_c2ad30.js"
+  ],
+  "source": "entry"
+});
